requirements.txt
# Backend
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.1

# AI & RAG Ecosystem
langchain
langchain-community
langchain-google-genai
langchain-core
langchain-chroma
langgraph
langgraph-checkpoint-sqlite

# Vector Database
chromadb>=0.4.24

# Utilities
tiktoken>=0.6.0
pandas>=2.2.0
httpx>=0.27.0
urllib3<2.0

# Frontend
streamlit>=1.30.0

# Testing
pytest
